refactor: 迁移音频处理 Worker
