+++ xl_updata_tool/app/features/audio/worker.py
@@ -1,12 +1,12 @@
 """Audio Feature 的后台任务入口。
 
-P1 先把 Worker 的归属和调用入口固定在功能域；具体 bank/debank 业务仍复用
-迁移期实现，待后续拆成 Service/Adapter 后再移除这个兼容包装。
+Worker 只负责 QThread 生命周期、取消和信号映射；文件处理、分类、缓存和
+debank 调用全部由 Qt-free ``AudioDecryptProcessor`` 执行。
 """
 
 from PySide6.QtCore import QThread, Signal
 
-from app.ui.workers.audio_decrypt import AudioDecryptWorker as _LegacyAudioDecryptWorker
+from app.features.audio.processing import AudioDecryptProcessor
 
 
 class AudioCatalogWorker(QThread):
@@ -27,8 +27,43 @@ class AudioCatalogWorker(QThread):
             self.failed.emit(str(error))
 
 
-class AudioDecryptWorker(_LegacyAudioDecryptWorker):
-    """功能域侧的音频解密 Worker 类型。"""
+class AudioDecryptWorker(QThread):
+    """将 AudioDecryptProcessor 映射到现有 AudioController 信号契约。"""
+
+    progress = Signal(str)
+    progress_value = Signal(int, int, str)
+    finished_decrypt = Signal()
+    cancelled_decrypt = Signal()
+    error = Signal(str)
+
+    def __init__(self, material_dir, audio_output_dir, debank_dir, force=False,
+                 lua_output_dir=None, parent=None):
+        super().__init__(parent)
+        self.processor = AudioDecryptProcessor(
+            material_dir,
+            audio_output_dir,
+            debank_dir,
+            force=force,
+            lua_output_dir=lua_output_dir,
+            progress_callback=self.progress.emit,
+            progress_value_callback=self.progress_value.emit,
+            cancel_check=lambda: self.isInterruptionRequested(),
+        )
+
+    def cancel(self):
+        self.requestInterruption()
+        self.processor.cancel()
+
+    def run(self):
+        try:
+            result = self.processor.process()
+        except Exception as error:
+            self.error.emit(str(error))
+            return
+        if result and result.get("cancelled"):
+            self.cancelled_decrypt.emit()
+        else:
+            self.finished_decrypt.emit()
 
 
 __all__ = ["AudioCatalogWorker", "AudioDecryptWorker"]
