+++ xl_updata_tool/docs/changes/issue-50-architecture-finalization.md
@@ -1,7 +1,9 @@
 # Issue #50：架构迁移收口
 
-状态：执行中，P0-P3 已完成，P4 待进入。
+状态：执行中，P0-P4 已完成，P5 待进入。
 
 本变更将沿现有 Feature Ownership 方向完成 Composition Root、兼容桥、Worker、领域代码归属、Character Parser 和终态架构测试收口。实施保持用户行为、输出目录、数据库、日志、信号、取消和失败语义不变。
 
 详细阶段计划、调查证据和验证结果见 `docs/worklog/issue-50-architecture-finalization/`。
+
+本阶段新增 Audio Feature 的 Qt-free 处理器和 Worker 适配器；旧音频 Worker 路径仅保留兼容转发，未改变音频输出、分类、增量缓存、取消和失败信号契约。
