+++ xl_updata_tool/docs/worklog/issue-50-architecture-finalization/progress.md
@@ -11,7 +11,8 @@
 - [completed] P1：PreviewPage、PreviewController 和导出编排已收回 Feature ownership；页面行为保护网、全量门禁和 Qt smoke 通过。
 - [completed] P2：五个 Feature 工厂、opaque 状态/进度/角标端口和 Qt-free Runtime registry 已建立；生产装配暂未切换。
 - [completed] P3：生产入口、descriptor 导航和 ImportResult 后处理 workflow 已切换；全量门禁和 runtime Qt smoke 通过。
-- [pending] P4：迁移 Audio 处理 Worker，去除 Feature 对旧 `app/ui/workers/audio_decrypt.py` 的生产依赖。
+- [completed] P4：迁移 Audio 处理 Worker；Qt-free 音频处理进入 Feature，生产 Feature Worker 不再依赖旧 `app/ui/workers/audio_decrypt.py`，旧路径保留薄兼容门面。
+- [pending] P5：迁移 Importer / AssetStudio Worker，去除 Feature 对旧导入 Worker 的生产依赖。
 
 ## 偏差
