+++ xl_updata_tool/docs/代码所有权与边界.md
@@ -6,7 +6,7 @@
 
 | 功能域 | 目标目录 | 当前迁移入口 | 普通功能改动的边界 |
 |---|---|---|---|
-| 音频 | `app/features/audio/` | `page.py`、`controller.py`、`service.py`、`tree.py`、`worker.py`；兼容入口为 `app/core/audio_*`、`app/core/album_map.py`、`app/ui/workers/audio_decrypt.py` | 音频页面、分类、审计、debank、播放器；不直接改 Shell |
+| 音频 | `app/features/audio/` | `page.py`、`controller.py`、`service.py`、`tree.py`、`processing.py`、`worker.py`；兼容入口为 `app/core/audio_*`、`app/core/album_map.py`、`app/ui/workers/audio_decrypt.py` | 音频页面、分类、审计、debank、播放器；不直接改 Shell |
 | 角色 | `app/features/characters/` | `page.py`、`controller.py`、`service.py`；兼容入口为 `app/core/character_*`、`app/ui/views/character_view.py` | Lua 解析、角色仓库、Wiki 页面、未读和 CSV；不直接改音频或版本 |
 | 预览 | `app/features/preview/` | `page.py`、`controller.py`、`service.py`、`item.py`、`worker.py`；兼容入口为 `app/core/preview_catalog.py`、`app/ui/features/fgui_atlas.py`、Spine 适配器 | 图片、FGUI、Spine、视频导出 |
 | 版本 | `app/features/versions/` | `page.py`、`controller.py`、`service.py`、`worker.py`；兼容入口为 `app/core/version_*`、`app/ui/workers/download.py` | 版本列表、检查更新、下载、删除和 Bundle 状态 |
@@ -20,7 +20,7 @@
 1. 新功能优先进入对应功能域，不继续向 `main_window.py` 添加领域状态和业务流程。
 2. Shell 不得直接导入角色解析器、音频仓库、数据库、AssetStudio、debank 或其他 Feature 内部实现。
 3. 新 Page 自己拥有控件，只暴露语义信号和展示状态方法，不调用 `parent._xxx`，不返回 `controls_dict`。
-4. 应用服务和领域逻辑不得依赖 PySide6；Qt Worker 只负责线程生命周期、取消、进度和结果信号。
+4. 应用服务和领域逻辑不得依赖 PySide6；Qt Worker 只负责线程生命周期、取消、进度和结果信号。Audio 的 `processing.py` 是该规则的处理器示例，旧 `app/ui/workers/audio_decrypt.py` 只保留兼容转发。
 5. Feature 之间通过契约、结果对象和显式后处理器通信，不互相导入内部实现。
 6. `app/core` 和旧 `app/ui` 在迁移期作为兼容层，不接受与迁移无关的新业务堆叠。
 7. `shell`、`bootstrap`、`platform`、`shared` 是公共架构区，改动需要单独 Issue、影响说明和额外审查。
